stage1: use RootDirectory to chroot instead of diagexec
systemd has the option "RootDirectory" to start an app in a chroot, so
we use it. Then, diagexec does not need to chroot to start an app, so it
can run as an unprivileged user. This fixes the options "user" and
"group" in the ACI manifest:
https://github.com/coreos/rocket/issues/539

diagexec is still used to diagnose the binary when the execution fails.

Since diagexec is a stage1 tool, it is not in the app's chroot
directory. To be able to execute it, diagexec is bind mounted to
/.diagexec in stage2.

Since the rkt environment file is outside the chroot, diagexec cannot
use it to setup the environment. Instead, the "EnvironmentFile" systemd
option is used. diagexec still has the code to setup the environment: it
is used for "rkt enter". Systemd's EnvironmentFile needs variables to be
separated by '\n' instead of '\0'. Rocket is updated to generate the
environment file accordingly.