stage1: create /etc/passwd with systemd-sysusers
Stage1 now creates a configuration files in /usr/lib/sysusers.d/ for
each application if the application specifies a non-root user in the
image manifest:

    "app": {
        "user": "1000",
        "group": "1000",

With this patch, "user" and "group" in the image manifest can only
contain uid and gid (not username and groupname).

A new systemd unit file "sysusers" calls systemd-sysusers before the app
is started. This will read the /usr/lib/sysusers.d/ configuration files
and generate /etc/passwd and /etc/group. With the example above, the
username and groupname will be "u1000".

systemd-sysusers is already available in stage1. systemd requires the
libnss_files library in order to open /etc/passwd, so it needs to be
added in stage1.

The systemd unit file of the app contain the username and groupname that
was created previously.